backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymssql
python-dotenv
pymemcache
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
xlsxwriter
pyreportjasper
ldap3